# services/gateway/docs/api/swagger.swagger.yaml
swagger: "2.0"
info:
  title: delivery/delivery_messages.proto
  version: version not set
...
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /v1/orders:
    get:
      operationId: OrderService_ListOrders
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/orderListOrdersResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: user_id
          in: query
          required: false
          type: string
        - name: status
          in: query
          required: false
          type: string
        - name: pagination.page
          in: query
          required: false
          type: integer
          format: int32
        - name: pagination.limit
          in: query
          required: false
          type: integer
          format: int32
        - name: pagination.total
          in: query
          required: false
          type: integer
          format: int32
      tags:
        - OrderService
    post:
      operationId: OrderService_CreateOrder
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/orderCreateOrderResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/orderCreateOrderRequest'
      tags:
        - OrderService
  /v1/orders/{order_id}:
    get:
      operationId: OrderService_GetOrder
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/orderGetOrderResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: order_id
          in: path
          required: true
          type: string
      tags:
        - OrderService
...
  /v1/orders/{order_id}/status:
    post:
      operationId: OrderService_UpdateOrderStatus
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/commonEmpty'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: order_id
          in: path
          required: true
          type: string
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/OrderServiceUpdateOrderStatusBody'
      tags:
        - OrderService
...
  /v1/users/login:
    post:
      operationId: UserService_LoginUser
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/userLoginUserResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/userLoginUserRequest'
      tags:
        - UserService
  /v1/users/register:
    post:
      operationId: UserService_RegisterUser
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/userRegisterUserResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/userRegisterUserRequest'
      tags:
        - UserService
...
  /v1/users/{user_id}/addresses:
    get:
      operationId: UserService_ListAddresses
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/userListAddressesResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: user_id
          in: path
          required: true
          type: string
      tags:
        - UserService
    post:
      operationId: UserService_AddAddress
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/userAddAddressResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: user_id
          in: path
          required: true
          type: string
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/UserServiceAddAddressBody'
      tags:
        - UserService
...
definitions:
...
  OrderServiceUpdateOrderStatusBody:
    type: object
    properties:
      status:
        $ref: '#/definitions/orderOrderStatus'
      reason:
        type: string
...
  UserServiceAddAddressBody:
    type: object
    properties:
      name:
        type: string
      phone:
        type: string
      postal_code:
        type: string
      prefecture:
        type: string
      city:
        type: string
      address_line1:
        type: string
      address_line2:
        type: string
      is_default:
        type: boolean
...
  commonEmpty:
    type: object
  commonMoney:
    type: object
    properties:
      currency:
        type: string
      units:
        type: string
        format: int64
      nanos:
        type: integer
        format: int32
  commonPagination:
    type: object
    properties:
      page:
        type: integer
        format: int32
      limit:
        type: integer
        format: int32
      total:
        type: integer
        format: int32
...
  orderCreateOrderItem:
    type: object
    properties:
      product_id:
        type: string
      variant_id:
        type: string
      quantity:
        type: integer
        format: int32
  orderCreateOrderRequest:
    type: object
    properties:
      user_id:
        type: string
      items:
        type: array
        items:
          type: object
          $ref: '#/definitions/orderCreateOrderItem'
      shipping_address:
        $ref: '#/definitions/orderShippingAddress'
      payment_method:
        $ref: '#/definitions/shinkansenorderPaymentMethod'
      points_to_apply:
        type: string
        format: int64
      delivery_slot_id:
        type: string
  orderCreateOrderResponse:
    type: object
    properties:
      order_id:
        type: string
      order_number:
        type: string
      status:
        $ref: '#/definitions/orderOrderStatus'
  orderGetOrderResponse:
    type: object
    properties:
      order:
        $ref: '#/definitions/orderOrder'
  orderListOrdersResponse:
    type: object
    properties:
      orders:
        type: array
        items:
          type: object
          $ref: '#/definitions/orderOrder'
      pagination:
        $ref: '#/definitions/commonPagination'
  orderOrder:
    type: object
    properties:
      id:
        type: string
      order_number:
        type: string
      user_id:
        type: string
      status:
        $ref: '#/definitions/orderOrderStatus'
      subtotal_amount:
        $ref: '#/definitions/commonMoney'
      tax_amount:
        $ref: '#/definitions/commonMoney'
      discount_amount:
        $ref: '#/definitions/commonMoney'
      total_amount:
        $ref: '#/definitions/commonMoney'
      points_applied:
        type: string
        format: int64
      shipping_address:
        $ref: '#/definitions/orderShippingAddress'
      payment_method:
        $ref: '#/definitions/shinkansenorderPaymentMethod'
      created_at:
        type: string
        format: date-time
      updated_at:
        type: string
        format: date-time
      delivery_slot_id:
        type: string
      estimated_delivery_at:
        type: string
        format: date-time
      items:
        type: array
        items:
          type: object
          $ref: '#/definitions/orderOrderItem'
  orderOrderItem:
    type: object
    properties:
      id:
        type: string
      product_id:
        type: string
      variant_id:
        type: string
      product_name:
        type: string
      quantity:
        type: integer
        format: int32
      unit_price:
        $ref: '#/definitions/commonMoney'
      total_price:
        $ref: '#/definitions/commonMoney'
...
  orderOrderStatus:
    type: string
    enum:
      - ORDER_STATUS_UNSPECIFIED
      - ORDER_STATUS_PENDING
      - ORDER_STATUS_CONFIRMED
      - ORDER_STATUS_PROCESSING
      - ORDER_STATUS_SHIPPED
      - ORDER_STATUS_IN_TRANSIT
      - ORDER_STATUS_DELIVERED
      - ORDER_STATUS_CANCELLED
      - ORDER_STATUS_EXPIRED
      - ORDER_STATUS_READY_FOR_PICKUP
      - ORDER_STATUS_PICKED_UP
      - ORDER_STATUS_FAILED_DELIVERY
      - ORDER_STATUS_RETURNED
    default: ORDER_STATUS_UNSPECIFIED
  orderShippingAddress:
    type: object
    properties:
      name:
        type: string
      phone:
        type: string
      postal_code:
        type: string
      prefecture:
        type: string
      city:
        type: string
      address_line1:
        type: string
      address_line2:
        type: string
...
  protobufAny:
    type: object
    properties:
      '@type':
        type: string
    additionalProperties: {}
  rpcStatus:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      details:
        type: array
        items:
          type: object
          $ref: '#/definitions/protobufAny'
...
  shinkansenorderPaymentMethod:
    type: string
    enum:
      - PAYMENT_METHOD_UNSPECIFIED
      - PAYMENT_METHOD_CREDIT_CARD
      - PAYMENT_METHOD_KONBINI_SEVENELEVEN
      - PAYMENT_METHOD_KONBINI_LAWSON
      - PAYMENT_METHOD_KONBINI_FAMILYMART
      - PAYMENT_METHOD_PAYPAY
      - PAYMENT_METHOD_RAKUTEN_PAY
    default: PAYMENT_METHOD_UNSPECIFIED
...
  userAddAddressResponse:
    type: object
    properties:
      address_id:
        type: string
  userAddress:
    type: object
    properties:
      id:
        type: string
      user_id:
        type: string
      name:
        type: string
      phone:
        type: string
      postal_code:
        type: string
      prefecture:
        type: string
      city:
        type: string
      address_line1:
        type: string
      address_line2:
        type: string
      is_default:
        type: boolean
      created_at:
        type: string
        format: date-time
...
  userListAddressesResponse:
    type: object
    properties:
      addresses:
        type: array
        items:
          type: object
          $ref: '#/definitions/userAddress'
  userLoginUserRequest:
    type: object
    properties:
      email:
        type: string
      password:
        type: string
  userLoginUserResponse:
    type: object
    properties:
      user_id:
        type: string
      access_token:
        type: string
      refresh_token:
        type: string
  userRegisterUserRequest:
    type: object
    properties:
      email:
        type: string
      password:
        type: string
      name:
        type: string
      phone:
        type: string
  userRegisterUserResponse:
    type: object
    properties:
      user_id:
        type: string
      access_token:
        type: string
      refresh_token:
        type: string
...